backend/requirements.txt
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.2
mysql-connector-python==9.2.0